Separated the classes into their own individual files
